requirements.txt
streamlit==1.29.0
sqlalchemy==2.0.23
openai==1.55.3
httpx==0.27.2
Pillow==10.1.0
python-dotenv==1.0.0